Explicitly define which blocks receive global fields
