+++ wp-content/plugins/core/src/Blocks/Blocks_Subscriber.php
@@ -7,6 +7,7 @@ use Tribe\Libs\ACF\Block_Config;
 use Tribe\Libs\ACF\Block_Registrar;
 use Tribe\Libs\ACF\Block_Renderer;
 use Tribe\Libs\Container\Abstract_Subscriber;
+use Tribe\Project\Blocks\Global_Field_Meta\Block_Controller;
 
 class Blocks_Subscriber extends Abstract_Subscriber {
 
@@ -34,25 +35,25 @@ class Blocks_Subscriber extends Abstract_Subscriber {
 			}
 		}, 10, 0 );
 
-		$this->register_global_blocks();
+		$this->register_global_block_fields();
 	}
 
 	/**
 	 * @throws RuntimeException
 	 */
-	private function register_global_blocks(): void {
-		$bypass_checker = $this->container->get( Block_Bypass_Checker::class );
+	private function register_global_block_fields(): void {
+		add_filter( 'tribe/block/register/fields', function ( $fields, Block_Config $block ): array {
 
-		add_filter( 'tribe/block/register/fields', function ( $fields, Block_Config $block ) use ( $bypass_checker ): array {
+			$controller = $this->container->get( Block_Controller::class );
 
-			// Bypass adding global fields to specified blocks
-			if ( $bypass_checker->bypass( $block::NAME ) ) {
+			// Only blocks specified will have global fields added to them.
+			if ( ! $controller->allowed( $block::NAME ) ) {
 				return $fields;
 			}
 
-			/** @var Global_Field_Meta\Meta $meta */
+			/** @var Global_Field_Meta\Meta\Meta $meta */
 			foreach ( $this->container->get( Blocks_Definer::GLOBAL_BLOCK_FIELD_COLLECTION ) as $meta ) {
-				if ( ! $meta instanceof Global_Field_Meta\Meta ) {
+				if ( ! $meta instanceof Global_Field_Meta\Meta\Meta ) {
 					throw new RuntimeException(
 						sprintf(
 							'%s is not an instance of \Tribe\Project\Blocks\Global_Field_Meta\Meta',
