+++ wp-content/plugins/core/src/Blocks/Global_Field_Meta/Meta/Block_Meta.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Tribe\Project\Blocks\Global_Field_Meta;
+namespace Tribe\Project\Blocks\Global_Field_Meta\Meta;
 
 use InvalidArgumentException;
 use Tribe\Libs\ACF\Field;
